added carousel to plotlist
